requirements.txt
appdirs==1.4.3
mock==1.3.0
nose==1.3.7
packaging==16.8
pbr==2.0.0
pydash==3.4.8
pyparsing==2.2.0
PyYAML==3.12
six==1.10.0
